@@ -12,7 +12,6 @@ import androidx.appcompat.app.AppCompatActivity;
 public class LogoActivity extends AppCompatActivity {
 TextView textView;
 ImageView imageView;
-private Animation textAnim;//создаем переменные для анимации
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -20,7 +19,8 @@ private Animation textAnim;//создаем переменные для аним
         setContentView(R.layout.activity_logo);
         textView = findViewById(R.id.textLogo);
         imageView = findViewById(R.id.imageLogo);
-        textAnim = AnimationUtils.loadAnimation(getApplicationContext(),R.anim.anim_logo);// создаем анимацию
+        //создаем переменные для анимации
+        Animation textAnim = AnimationUtils.loadAnimation(getApplicationContext(), R.anim.anim_logo);// создаем анимацию
         imageView.startAnimation(textAnim); //запускаем анимацию
         textView.startAnimation(textAnim); //запускаем анимацию
         startMainActivity();
@@ -29,20 +29,16 @@ private Animation textAnim;//создаем переменные для аним
     private void startMainActivity(){
 
         //создаем отдельный поток для счетчика--------------------
-        new Thread(new Runnable(){
-            @Override
-            //чтобы отобразить изменения в основном потоке
-            public void run() {
-                //исключение
-                try {
-                    Thread.sleep(2000); //задержка в 2 сек
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-                Intent intent = new Intent(LogoActivity.this, MainActivity.class);
-                startActivity(intent);
+        //чтобы отобразить изменения в основном потоке
+        new Thread(() -> {
+            //исключение
+            try {
+                Thread.sleep(2000); //задержка в 2 сек
+            } catch (InterruptedException e) {
+                e.printStackTrace();
             }
-
+            Intent intent = new Intent(LogoActivity.this, MainActivity.class);
+            startActivity(intent);
         }).start();
         //конец отдельного потока-----------------------------------------------------------
 
